+++ src/app/(customer)/loading.tsx
@@ -1,0 +1,7 @@
+export default function CustomerLoading() {
+  return (
+    <div className="flex min-h-screen items-center justify-center bg-orange-50">
+      <div className="h-8 w-8 animate-spin rounded-full border-4 border-orange-300 border-t-orange-600" />
+    </div>
+  );
+}
