+++ src/app/(customer)/products/page.tsx
@@ -1,87 +1,18 @@
-"use client";
+import type { Metadata } from "next";
+import { getAvailableProducts } from "@/db/queries/products";
+import { ProductList } from "@/components/product-list";
 
-import { useEffect, useState } from "react";
-import { ProductCard } from "@/components/product-card";
-import type { Product } from "@/types";
+export const metadata: Metadata = {
+  title: "商品一覧",
+};
 
-export default function ProductsPage() {
-  const [products, setProducts] = useState<Product[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [toast, setToast] = useState<string | null>(null);
-
-  useEffect(() => {
-    fetch("/api/products")
-      .then((res) => {
-        if (!res.ok) throw new Error("商品の取得に失敗しました");
-        return res.json();
-      })
-      .then(setProducts)
-      .catch((e) => setError(e.message))
-      .finally(() => setLoading(false));
-  }, []);
-
-  function handleAddToCart(productId: string, quantity: number) {
-    const product = products.find((p) => p.id === productId);
-    if (!product) return;
-
-    const cart = JSON.parse(localStorage.getItem("cart") ?? "[]");
-    const existing = cart.find((item: { id: string }) => item.id === productId);
-    if (existing) {
-      existing.quantity += quantity;
-    } else {
-      cart.push({
-        id: product.id,
-        name: product.name,
-        priceJpy: product.priceJpy,
-        quantity,
-      });
-    }
-    localStorage.setItem("cart", JSON.stringify(cart));
-    window.dispatchEvent(new Event("cart-updated"));
-
-    setToast(`${product.name} を ${quantity}個 カートに追加しました`);
-    setTimeout(() => setToast(null), 2000);
-  }
+export default async function ProductsPage() {
+  const products = await getAvailableProducts();
 
   return (
     <div className="min-h-screen bg-orange-50 p-4">
       <h1 className="mb-6 text-2xl font-bold text-orange-600">商品一覧</h1>
-      {loading && (
-        <div className="flex justify-center py-12">
-          <div className="h-8 w-8 animate-spin rounded-full border-4 border-orange-300 border-t-orange-600" />
-        </div>
-      )}
-      {error && (
-        <div className="rounded-lg bg-red-50 p-4 text-center text-red-600">
-          <p>{error}</p>
-          <button
-            onClick={() => window.location.reload()}
-            className="mt-2 text-sm underline"
-          >
-            再読み込み
-          </button>
-        </div>
-      )}
-      {!loading && !error && (
-        <div className="grid gap-4 sm:grid-cols-2">
-          {products.map((product) => (
-            <ProductCard
-              key={product.id}
-              {...product}
-              onAddToCart={handleAddToCart}
-            />
-          ))}
-        </div>
-      )}
-      {!loading && !error && products.length === 0 && (
-        <p className="text-center text-gray-500">現在販売中の商品はありません</p>
-      )}
-      {toast && (
-        <div className="fixed bottom-6 left-1/2 -translate-x-1/2 rounded-full bg-gray-800 px-6 py-3 text-sm text-white shadow-lg">
-          {toast}
-        </div>
-      )}
+      <ProductList products={products} />
     </div>
   );
 }
